Rename PdoGateway\Gateway to PdoDataGateway\PdoDataGateway
